animate: make fire into a working animation
