+++ plugins/animate/meson.build
@@ -1,5 +1,6 @@
 animiate = shared_module('animate',
                          ['animate.cpp',
+                          'fire/particle.cpp',
                           'fire/fire.cpp'],
                          include_directories: [wayfire_api_inc, wayfire_conf_inc],
                          dependencies: [wlroots, pixman, wfconfig],
